app/requirements.txt
fastapi==0.88.0
uvicorn[standard]==0.20.0
ujson==5.5.0
python-dotenv==0.21.0
sqlmodel==0.0.8
asyncpg==0.27.0
alembic==1.8.1
sqlalchemy-utils==0.38.3
mypy==0.961
yesqa==1.3.0
black==22.6.0
isort==5.10.1
pytest==6.2.5
pytest-asyncio==0.19.0
requests==2.28.1
httpx==0.23.0
psycopg2==2.9.10
celery==5.4.0
redis==5.2.1
